combined annotator files
