+++ src/grafzahl/language/annotator.rs
@@ -1,4 +1,31 @@
-use crate::grafzahl::language::annotation::{Annotation, LineType};
+use std::fmt::{Display, Formatter};
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum LineType {
+    Comment,
+    Code,
+}
+
+impl Display for LineType {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            LineType::Comment => write!(f, "{}", "Comm"),
+            LineType::Code => write!(f, "CODE"),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct Annotation {
+    pub(crate) line: String,
+    pub(crate) kind: LineType,
+}
+
+impl Annotation {
+    pub fn is_comment(&self) -> bool {
+        self.kind == LineType::Comment
+    }
+}
 
 /// Annotates Lines with if they are a comment or not
 pub fn annotate(lines: Vec<String>, inline_symbols: &Vec<String>, block_start: &Vec<String>, block_end: &Vec<String>) -> Vec<Annotation> {
