+++ src/grafzahl/language/mod.rs
@@ -1,3 +1,2 @@
 pub mod languages;
-pub mod general_annotator;
-pub mod annotation;
+pub mod annotator;
